socket: macOS: add MPTCP support
Multipath TCP (MPTCP), standardized in RFC8684 [1], is a TCP extension
that enables a TCP connection to use different paths.

MPTCP on macOS should be used via Apple's Network API [2] which is quite
different from the socket API.

On the other hand, an undocumented way to use MPTCP is to create a
socket with the AF_MULTIPATH family, and then use Apple's connectx().

To avoid duplicating the code of connectx() for TCP Fast Open and MPTCP,
a new helper has been created: do_connectx().

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>